+++ doc/whatsnew.rst
@@ -15,7 +15,11 @@ v0.8.dev
 
 - Enhance :func:`pyriemann.utils.mean.mean_covariance` to support "power" and "poweuclid" metrics. :pr:`329` by :user:`qbarthelemy`
 
-- Add an example using fNIRS data with a new estimator called `BlockKernels` for classifying HbO and HbR signals. :pr:`323` by :user:`timnaher`
+- Add tangent space alignment (TSA) in transfer learning module.
+  ``TLStretch`` is deprecated and renamed into :class:`pyriemann.transfer._estimators.TLScale`,
+  and ``TSclassifier`` into :class:`pyriemann.classification.TSClassifier`. :pr:`320` by :user:`qbarthelemy`
+
+- Add an example using fNIRS data with a new estimator called `HybridBlocks` for classifying HbO and HbR signals. :pr:`323` by :user:`timnaher`
 
 v0.7 (October 2024)
 -------------------